[dependencies-update] change - rename `Carthage` -> `CarthageController`
